+++ backend/packages/gitlab/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cat-factory/gitlab
 
+## 0.6.2
+
+### Patch Changes
+
+- Updated dependencies [55661f4]
+  - @cat-factory/contracts@0.88.0
+  - @cat-factory/kernel@0.76.0
+
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [ca5c3e8]
+  - @cat-factory/contracts@0.87.0
+  - @cat-factory/kernel@0.75.0
+
 ## 0.6.0
 
 ### Minor Changes